set up firebase hosting
